fix NullPointer for SharedPreferences

+++ app/src/main/java/skranz/pnpmanager/Mob.kt
@@ -18,8 +18,8 @@ class Mob(context: Context) {
         set(value) = prefs.edit().putString(NAME, value).apply()
     var health: Int
         get() = prefs.getInt(HEALTH, 20)
-        set(value) = prefs.edit().putInt(HEALTH, value).apply()
+        set(value) = prefs.edit().putInt(HEALTH, value.toInt()).apply()
     var armor: Int
         get() = prefs.getInt(ARMOR, 0)
-        set(value) = prefs.edit().putInt(ARMOR, value).apply()
+        set(value) = prefs.edit().putInt(ARMOR, value.toInt()).apply()
 }
